+++ core/tauri-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## \[2.0.0-beta.9]
+
+### Dependencies
+
+- Upgraded to `tauri-utils@2.0.0-beta.9`
+
 ## \[2.0.0-beta.8]
 
 ### Dependencies